Rename cli folder to basic-cli to try passing CI
