+++ src/index.ts
@@ -1,5 +1,5 @@
 export { overrideArgs, TooManyArguments, UnknownOptionError, Parsable } from './argv-parser';
-export { Cli, CliContext, CliOption } from './cli';
+export { Cli, CliContext, CliOption } from './basic-cli';
 export { CliCommand, createCliCommand, OptionNameNotUnique } from './cli-command';
 export * from './commands';
 export * from './plugin-cli';
